requirements.txt
accelerate>=0.34.0
bitsandbytes>=0.43.0
numpy>=1.26.0
openai>=1.0.0
opencv-python>=4.9.0
Pillow>=10.0.0
torch>=2.5.0
tqdm>=4.66.0
transformers>=4.51.0
unsloth>=2025.4.0
